update IPlugOSCEditor for windows
