+++ Examples/IPlugOSCEditor/projects/IPlugOSCEditor-aax.vcxproj
@@ -443,12 +443,18 @@
     <ClCompile Include="..\..\..\IPlug\AAX\IPlugAAX_Parameters.cpp" />
     <ClCompile Include="..\..\..\IPlug\AAX\IPlugAAX_Describe.cpp" />
     <ClCompile Include="..\..\..\IPlug\AAX\IPlugAAX.cpp" />
+    <ClCompile Include="..\..\..\IPlug\Extras\OSC\IPlugOSC.cpp" />
+    <ClCompile Include="..\..\..\IPlug\Extras\OSC\IPlugOSC_msg.cpp" />
     <ClCompile Include="..\..\..\IPlug\IPlugAPIBase.cpp" />
     <ClCompile Include="..\..\..\IPlug\IPlugParameter.cpp" />
     <ClCompile Include="..\..\..\IPlug\IPlugPaths.cpp" />
     <ClCompile Include="..\..\..\IPlug\IPlugPluginBase.cpp" />
     <ClCompile Include="..\..\..\IPlug\IPlugProcessor.cpp" />
     <ClCompile Include="..\..\..\IPlug\IPlugTimer.cpp" />
+    <ClCompile Include="..\..\..\WDL\jnetlib\asyncdns.cpp" />
+    <ClCompile Include="..\..\..\WDL\jnetlib\connection.cpp" />
+    <ClCompile Include="..\..\..\WDL\jnetlib\listen.cpp" />
+    <ClCompile Include="..\..\..\WDL\jnetlib\util.cpp" />
     <ClCompile Include="..\IPlugOSCEditor.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -518,6 +524,8 @@
     <ClInclude Include="..\..\..\IPlug\AAX\IPlugAAX_Parameters.h" />
     <ClInclude Include="..\..\..\IPlug\AAX\IPlugAAX_TaperDelegate.h" />
     <ClInclude Include="..\..\..\IPlug\AAX\IPlugAAX.h" />
+    <ClInclude Include="..\..\..\IPlug\Extras\OSC\IPlugOSC.h" />
+    <ClInclude Include="..\..\..\IPlug\Extras\OSC\IPlugOSC_msg.h" />
     <ClInclude Include="..\..\..\IPlug\IPlugAPIBase.h" />
     <ClInclude Include="..\..\..\IPlug\IPlugConstants.h" />
     <ClInclude Include="..\..\..\IPlug\IPlugDelegate_select.h" />
@@ -535,6 +543,13 @@
     <ClInclude Include="..\..\..\IPlug\IPlugUtilities.h" />
     <ClInclude Include="..\..\..\IPlug\IPlug_include_in_plug_hdr.h" />
     <ClInclude Include="..\..\..\IPlug\IPlug_include_in_plug_src.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\asyncdns.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\connection.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\irc_util.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\jnetlib.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\listen.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\netinc.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\util.h" />
     <ClInclude Include="..\config.h" />
     <ClInclude Include="..\IPlugOSCEditor.h" />
     <ClInclude Include="..\resources\resource.h" />
