+++ Examples/IPlugOSCEditor/projects/IPlugOSCEditor-aax.vcxproj.filters
@@ -74,6 +74,24 @@
     <ClCompile Include="..\..\..\IGraphics\Drawing\IGraphicsSkia.cpp">
       <Filter>IGraphics\Drawing</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\..\WDL\jnetlib\asyncdns.cpp">
+      <Filter>WDL\jnetlib</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\..\WDL\jnetlib\connection.cpp">
+      <Filter>WDL\jnetlib</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\..\WDL\jnetlib\listen.cpp">
+      <Filter>WDL\jnetlib</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\..\WDL\jnetlib\util.cpp">
+      <Filter>WDL\jnetlib</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\..\IPlug\Extras\OSC\IPlugOSC.cpp">
+      <Filter>IPlug\Extras\OSC</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\..\IPlug\Extras\OSC\IPlugOSC_msg.cpp">
+      <Filter>IPlug\Extras\OSC</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\IPlugOSCEditor.h" />
@@ -243,6 +261,33 @@
     <ClInclude Include="..\..\..\IGraphics\Drawing\IGraphicsSkia.h">
       <Filter>IGraphics\Drawing</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\..\WDL\jnetlib\asyncdns.h">
+      <Filter>WDL\jnetlib</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\WDL\jnetlib\connection.h">
+      <Filter>WDL\jnetlib</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\WDL\jnetlib\irc_util.h">
+      <Filter>WDL\jnetlib</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\WDL\jnetlib\jnetlib.h">
+      <Filter>WDL\jnetlib</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\WDL\jnetlib\listen.h">
+      <Filter>WDL\jnetlib</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\WDL\jnetlib\netinc.h">
+      <Filter>WDL\jnetlib</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\WDL\jnetlib\util.h">
+      <Filter>WDL\jnetlib</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\IPlug\Extras\OSC\IPlugOSC.h">
+      <Filter>IPlug\Extras\OSC</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\IPlug\Extras\OSC\IPlugOSC_msg.h">
+      <Filter>IPlug\Extras\OSC</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="resources">
@@ -268,6 +313,18 @@
     </Filter>
     <Filter Include="IGraphics\Controls">
       <UniqueIdentifier>{3c740de7-2371-4065-bf3d-5af9193c7b7d}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="WDL">
+      <UniqueIdentifier>{c3aba9d3-a52e-4913-9ead-68f666e0ec96}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="WDL\jnetlib">
+      <UniqueIdentifier>{449c6428-e14e-47d0-b8ba-78f64fc770d2}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="IPlug\Extras">
+      <UniqueIdentifier>{f983ef11-dc17-44f7-9944-870c54838d75}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="IPlug\Extras\OSC">
+      <UniqueIdentifier>{ccf5e42f-eb63-4398-bb6d-9d258bbc3ae2}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
