+++ Examples/IPlugOSCEditor/projects/IPlugOSCEditor-vst2.vcxproj
@@ -284,6 +284,8 @@
     <ClInclude Include="..\..\..\IGraphics\Platforms\IGraphicsMac_view.h" />
     <ClInclude Include="..\..\..\IGraphics\Platforms\IGraphicsWeb.h" />
     <ClInclude Include="..\..\..\IGraphics\Platforms\IGraphicsWin.h" />
+    <ClInclude Include="..\..\..\IPlug\Extras\OSC\IPlugOSC.h" />
+    <ClInclude Include="..\..\..\IPlug\Extras\OSC\IPlugOSC_msg.h" />
     <ClInclude Include="..\..\..\IPlug\IPlugAPIBase.h" />
     <ClInclude Include="..\..\..\IPlug\IPlugConstants.h" />
     <ClInclude Include="..\..\..\IPlug\IPlugEditorDelegate.h" />
@@ -300,6 +302,13 @@
     <ClInclude Include="..\..\..\IPlug\IPlug_include_in_plug_hdr.h" />
     <ClInclude Include="..\..\..\IPlug\IPlug_include_in_plug_src.h" />
     <ClInclude Include="..\..\..\IPlug\VST2\IPlugVST2.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\asyncdns.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\connection.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\irc_util.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\jnetlib.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\listen.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\netinc.h" />
+    <ClInclude Include="..\..\..\WDL\jnetlib\util.h" />
     <ClInclude Include="..\IPlugOSCEditor.h" />
     <ClInclude Include="..\resources\resource.h" />
   </ItemGroup>
@@ -367,6 +376,8 @@
     <ClCompile Include="..\..\..\IGraphics\IGraphics.cpp" />
     <ClCompile Include="..\..\..\IGraphics\IGraphicsEditorDelegate.cpp" />
     <ClCompile Include="..\..\..\IGraphics\Platforms\IGraphicsWin.cpp" />
+    <ClCompile Include="..\..\..\IPlug\Extras\OSC\IPlugOSC.cpp" />
+    <ClCompile Include="..\..\..\IPlug\Extras\OSC\IPlugOSC_msg.cpp" />
     <ClCompile Include="..\..\..\IPlug\IPlugAPIBase.cpp" />
     <ClCompile Include="..\..\..\IPlug\IPlugParameter.cpp" />
     <ClCompile Include="..\..\..\IPlug\IPlugPaths.cpp" />
@@ -374,6 +385,10 @@
     <ClCompile Include="..\..\..\IPlug\IPlugProcessor.cpp" />
     <ClCompile Include="..\..\..\IPlug\IPlugTimer.cpp" />
     <ClCompile Include="..\..\..\IPlug\VST2\IPlugVST2.cpp" />
+    <ClCompile Include="..\..\..\WDL\jnetlib\asyncdns.cpp" />
+    <ClCompile Include="..\..\..\WDL\jnetlib\connection.cpp" />
+    <ClCompile Include="..\..\..\WDL\jnetlib\listen.cpp" />
+    <ClCompile Include="..\..\..\WDL\jnetlib\util.cpp" />
     <ClCompile Include="..\IPlugOSCEditor.cpp" />
   </ItemGroup>
   <ItemGroup>
